api_yamdb/requirements.txt
asgiref==3.2.10
attrs==21.2.0
certifi==2021.10.8
charset-normalizer==2.0.9
Django==2.2.16
django-filter==21.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
gunicorn==20.0.4
idna==3.3
importlib-metadata==4.10.0
iniconfig==1.1.1
packaging==21.3
pluggy==0.13.1
py==1.11.0
PyJWT==2.3.0
pyparsing==3.0.6
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
pytz==2021.3
python-dotenv==0.19.1
psycopg2-binary==2.8.6
requests==2.26.0
sqlparse==0.4.2
toml==0.10.2
typing_extensions==4.0.1
urllib3==1.26.7
zipp==3.6.0
